+++ Runtime/MXTimeline/MXClipBehaviour.cs
@@ -1,18 +1,25 @@
 using UnityEngine.Playables;
 
+#if UNITY_EDITOR
+using UnityEditor.Timeline;
+#endif
+
 namespace Voxell.MotionGFX
 {
   public class MXClipBehaviour : PlayableBehaviour
   {
     public MXScene scene;
 
-    // public override void ProcessFrame(Playable playable, FrameData info, object playerData)
-    // {
-    //   if (scene == null) return;
+    public void Init()
+    {
+      if (scene == null) return;
 
-    //   float playableTime = (float) playable.GetTime();
-    //   ISeqHolder seqHolder = scene as ISeqHolder;
-    //   seqHolder.Evaluate(playableTime);
-    // }
+      #if UNITY_EDITOR
+      float directorTime = (float) TimelineEditor.inspectedDirector.time;
+
+      ISeqHolder seqHolder = scene;
+      seqHolder.InitEvaluation(directorTime);
+      #endif
+    }
   }
 }