+++ Runtime/MXTimeline/MXTrack.cs
@@ -9,7 +9,7 @@ namespace Voxell.MotionGFX
   {
     public override Playable CreateTrackMixer(PlayableGraph graph, GameObject go, int inputCount)
     {
-      return base.CreateTrackMixer(graph, go, inputCount);
+      return ScriptPlayable<MXTrackMixer>.Create(graph, inputCount);
     }
   }
 }