+++ Runtime/MXScene/ISeqHolder.cs
@@ -19,7 +19,7 @@ namespace Voxell.MotionGFX
     {
       globalTime = math.clamp(globalTime, StartTime, EndTime);
       UpdatePreviousData(globalTime);
-      if (Holders?.Count == 0) return;
+      if (Holders == null || Holders.Count == 0) return;
 
       float localTime = globalTime - StartTime;
 
@@ -42,7 +42,7 @@ namespace Voxell.MotionGFX
     void IHolder.Evaluate(float globalTime)
     {
       globalTime = math.clamp(globalTime, StartTime, EndTime);
-      if (Holders?.Count == 0) return;
+      if (Holders == null || Holders?.Count == 0) return;
 
       float durationChange = Duration - PrevDuration;
       float localTime = globalTime - StartTime;