+++ CHANGELOG.md
@@ -1,9 +1,16 @@
+## [0.2.2]
+
+### Bug Fixes
+
+- Fixed `Holders` is null error when evalulating.
+
 ## [0.2.1]
 
 ### New Features
 
 - Added "set" type actions.
 - Added a whole lot of easing transitions.
+- Added `MXTimelineManager` to auto rebuild `PlayableDirector` graph when Unity reloads.
 
 ### Changes
 
@@ -12,7 +19,6 @@
 - Added "alpha" to package version.
 - Clamp global time to start and end in `ISeqHolder`.
 - Scene evalulation now happens at `MXTrackMixer`.
-- Added `MXTimelineManager` to auto rebuild `PlayableDirector` graph when Unity reloads.
 - Ignore evalulation for `ISeqHolder` when `Holders` list is null or has 0 count.
 - Create instance of `MXClipBehaviour` before creating playable out of it.